Team Project/Internet Technology/requirements.txt
dj-database-url==0.4.2
Django==1.11.6
django-bootstrap3==9.1.0
gunicorn==19.7.1
html5lib==0.999999999
numpy==1.13.3
pipenv==9.0.3
Pillow==4.2.1
python-dateutil==2.6.1
virtualenv==15.1.0
virtualenv-clone==0.2.6
whitenoise==3.3.1
django-mptt==0.9.0
psycopg2==2.7.3.2